fix some warnings/errors: * Some platforms want sys/wait.h * Someone forgot a newline in .h file * Cast pointer arithmetics to int, less warnings * Return 0 at the end of main() :)
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@12317 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ udpprofile/export.c
@@ -64,7 +64,7 @@ void dumpData(FILE *fd) {
 				"<cputime total=\"%lf\" totalsq=\"%lf\" />\n" \
 				"<realtime total=\"%lf\" totalsq=\"%lf\" />\n" \
 				"</stats></event>\n",
-				key.size - ((void *)p-(void *)key.data),p,
+				(int)(key.size - ((void *)p-(void *)key.data)),p,
 				entry->pf_count, entry->pf_cpu, entry->pf_cpu_sq,
 				entry->pf_real, entry->pf_real_sq);
